Move TimeRangeSelector out of analysisPageRows check. Move NothingHereText into its own file. Use NothingHereText if there are no rows to show.
